Seperate mouse listener from `main.rb` and introduce new directory structure
